first cve patchset to test 3.4 kernel
A first set of cve patches for 3.4 kernel for testing.

The list is not complete, and not curated into general/3.4 specific.

Extracted from kernel_lenovo_msm8974 based on cve topic
(chose this repo as it has cve topics set)

I left LineageOS gerrit changeid, project, and part of subject
in filename, can be removedif needed.