patches: Import more Linux 3.10 CVE patches
CVE patchsets for testing only on 3.10 kernels

The patch sets were created automated by filtering
LineageOS gerrit changes with a CVE-* topic
on cm-14.1 branches of the android_kernel_*
projects of version 3.10.

cve changes without a cve topic are not included.

This is a WIP, use only for testing

For reference: the kernels used for this version are:

    android_kernel_asus_P024
    android_kernel_bq_msm8939
    android_kernel_cyanogen_msm8916
    android_kernel_huawei_kiwi
    android_kernel_lge_msm8992
    android_kernel_motorola_msm8916
    android_kernel_motorola_msm8992
    android_kernel_samsung_apq8084
    android_kernel_samsung_msm8976

    android_kernel_oneplus_msm8996
    android_kernel_xiaomi_msm8996